quant-evolve/requirements.txt
# Core Dependencies for QuantEvolve
# Python Version: 3.9 or 3.10 (Zipline Reloaded compatibility)

# Backtesting Engine
zipline-reloaded>=2.4.0

# Performance Analytics
quantstats>=0.0.60

# Data Manipulation
pandas>=2.0.0
numpy>=1.24.0

# Statistical Functions
scipy>=1.10.0

# Configuration Parsing
pyyaml>=6.0

# Strategy Serialization
dataclasses-json>=0.6.0

# LLM Integration (choose one)
# Option 1: Local inference with vLLM
vllm>=0.3.0
# Option 2: OpenAI-compatible API
openai>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Development & Testing
pytest>=7.4.0
black>=23.0
flake8>=6.0

# Additional Utilities
tqdm>=4.65.0
joblib>=1.3.0
